cleanup

+++ modules/cli/src/main/java/builders/loom/cli/Loom.java
@@ -19,7 +19,6 @@ package builders.loom.cli;
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.channels.FileLock;
-import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -250,11 +249,17 @@ public final class Loom {
 
         try {
             Files.walkFileTree(loomDir, new SimpleFileVisitor<>() {
+
+                private boolean deleteDir;
+
                 @Override
-                public FileVisitResult preVisitDirectory(final Path dir, final BasicFileAttributes attrs) throws IOException {
+                public FileVisitResult preVisitDirectory(final Path dir,
+                                                         final BasicFileAttributes attrs) {
                     if (LoomPaths.configDir(projectBaseDir).equals(dir)) {
+                        deleteDir = false;
                         return FileVisitResult.SKIP_SUBTREE;
                     }
+                    deleteDir = true;
                     return FileVisitResult.CONTINUE;
                 }
 
@@ -268,10 +273,11 @@ public final class Loom {
                 @Override
                 public FileVisitResult postVisitDirectory(final Path dir, final IOException exc)
                     throws IOException {
-                    try {
+                    if (exc != null) {
+                        throw exc;
+                    }
+                    if (deleteDir) {
                         Files.delete(dir);
-                    } catch (DirectoryNotEmptyException dne) {
-                        // ok
                     }
                     return FileVisitResult.CONTINUE;
                 }
